requirements.txt
alembic==1.11.1
amqp==5.1.1
anyio==3.7.0
asgiref==3.7.2
base58==2.1.1
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.1
Deprecated==1.2.14
dnspython==2.3.0
ecdsa==0.18.0
ECPy==1.2.5
email-validator==2.0.0.post2
fastapi==0.96.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.3.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pika==1.3.2
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pytest==7.3.1
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
raven==6.10.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
types-Deprecated==1.2.9.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
wrapt==1.15.0
xrpl-py==1.8.0b0
